rename dirs and create release
