+++ src/distributors/adoptium/adoptopenjdk-models.ts
